refactor: remove alteration from core
