+++ packages/cli/jest.config.ts
@@ -1,0 +1,1 @@
+export { default } from '@silverhand/jest-config';
