+++ packages/cli/package.json
@@ -6,6 +6,9 @@
   "homepage": "https://github.com/logto-io/logto#readme",
   "license": "MPL-2.0",
   "main": "lib/index.js",
+  "exports": {
+    ".": "./lib"
+  },
   "bin": {
     "logto": "bin/logto",
     "lg": "bin/logto"
@@ -20,11 +23,13 @@
   },
   "scripts": {
     "precommit": "lint-staged",
-    "build": "rimraf lib && tsc",
+    "build": "rimraf lib && tsc -p tsconfig.build.json",
     "start": "node .",
     "start:dev": "ts-node --files src/index.ts",
     "lint": "eslint --ext .ts src",
     "lint:report": "pnpm lint --format json --output-file report.json",
+    "test": "jest",
+    "test:ci": "jest",
     "prepack": "pnpm build"
   },
   "engines": {
@@ -56,15 +61,18 @@
   },
   "devDependencies": {
     "@silverhand/eslint-config": "1.0.0",
+    "@silverhand/jest-config": "1.0.0",
     "@silverhand/ts-config": "1.0.0",
     "@types/decompress": "^4.2.4",
     "@types/fs-extra": "^9.0.13",
     "@types/inquirer": "^8.2.1",
+    "@types/jest": "^28.1.6",
     "@types/node": "^16.0.0",
     "@types/semver": "^7.3.12",
     "@types/tar": "^6.1.2",
     "@types/yargs": "^17.0.13",
     "eslint": "^8.21.0",
+    "jest": "^28.1.3",
     "lint-staged": "^13.0.0",
     "prettier": "^2.7.1",
     "rimraf": "^3.0.2",
