+++ packages/cli/tsconfig.json
@@ -7,7 +7,8 @@
     "target": "es2022"
   },
   "include": [
-    "src"
+    "src",
+    "jest.config.ts"
   ],
   "exclude": ["**/alteration-scripts"]
 }
