+++ packages/cli/tsconfig.test.json
@@ -1,0 +1,6 @@
+{
+  "extends": "./tsconfig",
+  "compilerOptions": {
+    "allowJs": true
+  }
+}
