--- packages/core/src/alteration/constants.ts
@@ -1,3 +1,0 @@
-export const logtoConfigsTableFilePath = 'node_modules/@logto/schemas/tables/logto_configs.sql';
-export const alterationFilesDirectorySource = 'node_modules/@logto/schemas/alterations';
-export const alterationFilesDirectory = 'alterations/';
